+++ test-helpers/src/thread_unix.cpp
@@ -53,7 +53,7 @@ namespace micro_profiler
 			timespec t = {};
 
 			::clock_gettime(CLOCK_THREAD_CPUTIME_ID, &t);
-			return mt::milliseconds(static_cast<unsigned int>(t.tv_sec * 1000 + t.tv_nsec / 1000000));
+			return mt::milliseconds(t.tv_sec * 1000 + t.tv_nsec / 1000000);
 		}
 
 		bool this_thread::set_description(const wchar_t * /*description*/)
